requirements.txt
Django
dj-database-url
psycopg2
djangorestframework
requests